refactor html files and change them to fragments
