+++ RhinoshareApp/rhinoshareapp/src/main/java/de/hhu/rhinoshareapp/controller/MainpageController.java
@@ -25,6 +25,7 @@ public class MainpageController {
         ActualUserChecker.checkActualUser(model, p, userRepository);
         Iterable<Article> articles= articleRepository.findAll();
         model.addAttribute("articles", articles);
+        model.addAttribute("mainActive","active");
         return "Article/viewAll";
     }
