+++ RhinoshareApp/rhinoshareapp/src/main/resources/templates/Admin/admin_articlemanagement.html
@@ -1,48 +1,36 @@
 <!DOCTYPE html>
-<html lang="de" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>Artikelverwaltung - Rhinoshare</title>
-    <link rel="stylesheet" type="text/css" href="/css/stylesheet.css">
-    <link rel="icon" href="/img/RhinoLogoSmall.png">
-</head>
-<body>
-<img id="titleImageTest" src="/img/rhinoLogoBigAdmin.png" width="300" height="75">
-<ul class="navigationbar">
-    <li class="navbaritem"><a href="/admin">Überblick</a></li>
-    <li class="navbaritem">
-        <a href="/admin/usermanagement">Benutzerverwaltung</a>
-    </li>
-    <li class="navbaritem"><a class="active" href="/admin/articlemanagement">Artikelverwaltung</a></li>
-    <li class="navbaritem"><a href="/admin/lendingmanagement">Ausleihen</a></li>
-    <li class="navbaritem"><a href="/admin/conflicthandling">Konfliktlösung</a></li>
-    <li class="navbaritem"><a href="/">Hauptseite</a></li>
-    <li style="float: right" class="navbaritem">
-        <a href="/logout">Ausloggen</a>
-    </li>
-</ul>
-<table id="useroverview">
-    <th>ID</th>
-    <th>Artikel</th>
-    <th>Besitzer</th>
-    <th>Kaution</th>
-    <th>Mietkosten</th>
-    <th>Aktion</th>
-    <th:block th:each="article : ${articles}">
-        <tr>
-            <td th:text="${article.articleID}">...</td>
-            <td th:text="${article.name}">...</td>
-            <td th:text="${article.owner.getUsername()}">...</td>
-            <td th:text="${article.deposit}">...</td>
-            <td th:text="${article.rent}">...</td>
-            <td>
-                <a th:href="|/admin/articlemanagement/delete/${article.getArticleID()}|">
-                    <button class="button_std">Löschen</button>
-                </a>
-            </td>
-        </tr>
-    </th:block>
-</table>
+<html lang="de" xmlns:th="http://www.thymeleaf.org" layout:decorate="~{Admin/admin_overview}"
+      xmlns:layout="http://www.w3.org/1999/xhtml">
 
-</body>
+    <head>
+        <title>Artikelverwaltung - Rhinoshare</title>
+    </head>
+
+    <body>
+        <div class="container" layout:fragment="content">
+            <table id="useroverview">
+                <th>ID</th>
+                <th>Artikel</th>
+                <th>Besitzer</th>
+                <th>Kaution</th>
+                <th>Mietkosten</th>
+                <th>Aktion</th>
+                <th:block th:each="article : ${articles}">
+                    <tr>
+                        <td th:text="${article.articleID}">...</td>
+                        <td th:text="${article.name}">...</td>
+                        <td th:text="${article.owner.getUsername()}">...</td>
+                        <td th:text="${article.deposit}">...</td>
+                        <td th:text="${article.rent}">...</td>
+                        <td>
+                            <a th:href="|/admin/articlemanagement/delete/${article.getArticleID()}|">
+                                <button class="button_std">Löschen</button>
+                            </a>
+                        </td>
+                    </tr>
+                </th:block>
+            </table>
+        </div>
+    </body>
+
 </html>
