+++ RhinoshareApp/rhinoshareapp/src/main/resources/templates/Admin/admin_lendingmanagement.html
@@ -1,52 +1,40 @@
 <!DOCTYPE html>
-<html lang="de" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>Ausleihenverwaltung - Rhinoshare</title>
-    <link rel="stylesheet" type="text/css" href="/css/stylesheet.css">
-    <link rel="icon" href="/img/RhinoLogoSmall.png">
-</head>
-<body>
-<img id="titleImageTest" src="/img/rhinoLogoBigAdmin.png" width="300" height="75">
-<ul class="navigationbar">
-    <li class="navbaritem"><a href="/admin">Überblick</a></li>
-    <li class="navbaritem">
-        <a href="/admin/usermanagement">Benutzerverwaltung</a>
-    </li>
-    <li class="navbaritem"><a href="/admin/articlemanagement">Artikelverwaltung</a></li>
-    <li class="navbaritem"><a class="active" href="/admin/lendingmanagement">Ausleihen</a></li>
-    <li class="navbaritem"><a href="/admin/conflicthandling">Konfliktlösung</a></li>
-    <li class="navbaritem"><a href="/">Hauptseite</a></li>
-    <li style="float: right" class="navbaritem">
-        <a href="/logout">Ausloggen</a>
-    </li>
-</ul>
-<table id="useroverview">
-    <th>ID</th>
-    <th>Artikel</th>
-    <th>Startdatum</th>
-    <th>Enddatum</th>
-    <th>Verkiehen an</th>
-    <th>Besitzer</th>
-    <th>Bemerkung</th>
-    <th>Aktion</th>
-    <th:block th:each="lending : ${lendings}">
-        <tr>
-            <td th:text="${lending.lendingID}">...</td>
-            <td th:text="${lending.lendedArticle.getName()}">...</td>
-            <td th:text="${lending.formattedStartDate}">...</td>
-            <td th:text="${lending.formattedEndDate}">...</td>
-            <td th:text="${lending.lendingPerson.getUsername()}">...</td>
-            <td th:text="${lending.lendedArticle.getOwner().getUsername()}">...</td>
-            <td th:text="${lending.warning}">...</td>
-            <td>
-                <a th:href="|/admin/lendingmanagement/delete/${lending.getLendingID()}|">
-                    <button class="button_std">Löschen</button>
-                </a>
-            </td>
-        </tr>
-    </th:block>
-</table>
+<html lang="de" xmlns:th="http://www.thymeleaf.org" layout:decorate="~{Admin/admin_overview}"
+      xmlns:layout="http://www.w3.org/1999/xhtml">
 
-</body>
+    <head>
+        <title>Ausleihenverwaltung - Rhinoshare</title>
+    </head>
+
+    <body>
+        <div class="container" layout:fragment="content">
+            <table id="useroverview">
+                <th>ID</th>
+                <th>Artikel</th>
+                <th>Startdatum</th>
+                <th>Enddatum</th>
+                <th>Verkiehen an</th>
+                <th>Besitzer</th>
+                <th>Bemerkung</th>
+                <th>Aktion</th>
+                <th:block th:each="lending : ${lendings}">
+                    <tr>
+                        <td th:text="${lending.lendingID}">...</td>
+                        <td th:text="${lending.lendedArticle.getName()}">...</td>
+                        <td th:text="${lending.formattedStartDate}">...</td>
+                        <td th:text="${lending.formattedEndDate}">...</td>
+                        <td th:text="${lending.lendingPerson.getUsername()}">...</td>
+                        <td th:text="${lending.lendedArticle.getOwner().getUsername()}">...</td>
+                        <td th:text="${lending.warning}">...</td>
+                        <td>
+                            <a th:href="|/admin/lendingmanagement/delete/${lending.getLendingID()}|">
+                                <button class="button_std">Löschen</button>
+                            </a>
+                        </td>
+                    </tr>
+                </th:block>
+            </table>
+        </div>
+    </body>
+
 </html>
