+++ RhinoshareApp/rhinoshareapp/src/main/resources/templates/Admin/admin_overview.html
@@ -1,28 +1,31 @@
 <!DOCTYPE html>
-<html lang="de" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>Administration - Rhinoshare</title>
-    <link rel="stylesheet" type="text/css" href="/css/stylesheet.css">
-    <link rel="icon" href="/img/RhinoLogoSmall.png">
-</head>
-<body>
-<img id="titleImageTest" src="/img/rhinoLogoBigAdmin.png" width="300" height="75">
-<ul class="navigationbar">
-    <li class="navbaritem"><a class="active" href="/admin">Überblick</a></li>
-    <li class="navbaritem">
-        <a href="/admin/usermanagement">Benutzerverwaltung</a>
-    </li>
-    <li class="navbaritem"><a href="/admin/articlemanagement">Artikelverwaltung</a></li>
-    <li class="navbaritem"><a href="/admin/lendingmanagement">Ausleihen</a></li>
-    <li class="navbaritem"><a href="/admin/conflicthandling">Konfliktlösung</a></li>
-    <li class="navbaritem"><a href="/">Hauptseite</a></li>
-    <li style="float: right" class="navbaritem">
-        <a href="/logout">Ausloggen</a>
-    </li>
-</ul>
-</ul>
-<h1>ADMIN</h1>
+<html lang="de" xmlns:th="http://www.thymeleaf.org" xmlns:layout="http://www.w3.org/1999/xhtml">
 
-</body>
+    <head>
+        <meta charset="utf-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+        <title>Administration - Rhinoshare</title>
+        <link rel="stylesheet" type="text/css" href="/css/stylesheet.css">
+        <link rel="icon" href="/img/RhinoLogoSmall.png">
+        <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
+    </head>
+
+    <body>
+        <img id="titleImageTest" src="/img/rhinoLogoBigAdmin.png" width="300" height="75">
+        <ul class="navigationbar">
+            <li class="navbaritem" th:classappend="${mainActive}"><a href="/admin">Überblick</a></li>
+            <li class="navbaritem" th:classappend="${userActive}"><a href="/admin/usermanagement">Benutzerverwaltung</a></li>
+            <li class="navbaritem" th:classappend="${articleActive}"><a href="/admin/articlemanagement">Artikelverwaltung</a></li>
+            <li class="navbaritem" th:classappend="${lendingActive}"><a href="/admin/lendingmanagement">Ausleihen</a></li>
+            <li style="float: right" class="navbaritem"><a href="/logout">Ausloggen</a></li>
+            <li style="float: right" class="navbaritem"><a href="/">Zur Hauptseite</a></li>
+        </ul>
+    </body>
+
+    <div class="contentFragment">
+        <div layout:fragment="content">
+            Content goes here
+        </div>
+    </div>
+
 </html>
