+++ RhinoshareApp/rhinoshareapp/src/main/resources/templates/Admin/admin_userEdit.html
@@ -1,72 +1,61 @@
 <!DOCTYPE html>
-<html lang="de" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>Benutzerverwaltung - Rhinoshare</title>
-    <link rel="stylesheet" type="text/css" href="/css/stylesheet.css">
-    <link rel="icon" href="/img/RhinoLogoSmall.png">
-</head>
-<body>
-<img id="titleImageTest" src="/img/rhinoLogoBigAdmin.png" width="300" height="75">
-<ul class="navigationbar">
-    <li class="navbaritem"><a href="/admin">Überblick</a></li>
-    <li class="navbaritem">
-        <a class="active" href="/admin/usermanagement">Benutzerverwaltung</a>
-    </li>
-    <li class="navbaritem"><a href="/admin/articlemanagement">Artikelverwaltung</a></li>
-    <li class="navbaritem"><a href="/admin/lendingmanagement">Ausleihen</a></li>
-    <li class="navbaritem"><a href="/admin/conflicthandling">Konfliktlösung</a></li>
-    <li class="navbaritem"><a href="/">Hauptseite</a></li>
-    <li style="float: right" class="navbaritem">
-        <a href="/logout">Ausloggen</a>
-    </li>
-</ul>
-<div>
-    <a href="/admin/usermanagement">
-        <button class="button_std">Zurück</button>
-    </a>
-</div>
-<form method="post" style="text-align: center;">
-    <table class="tg">
-        <tr>
-            <td class="tg-lap0">Username:</td>
-            <td class="tg-lap0"><input th:name="username" type="text" th:placeholder="${user.username}"/></td>
-        </tr>
-        <tr>
-            <td class="tg-lap0">Nachname:</td>
-            <td class="tg-lap0"><input th:name="name" type="text" th:placeholder="${user.name}"/></td>
-        </tr>
-        <tr>
-            <td class="tg-lap0">Vorname:</td>
-            <td class="tg-lap0"><input th:name="surname" type="text" th:placeholder="${user.surname}"/></td>
-        </tr>
-        <tr>
-            <td class="tg-lap0">E-Mail Adresse:</td>
-            <td class="tg-lap0"><input th:name="email" type="email" th:placeholder="${user.email}"/></td>
-        </tr>
-        <tr>
-            <td class="tg-lap0">Straße:</td>
-            <td class="tg-lap0"><input th:name="street" type="text" th:placeholder="${user.address.street}"/></td>
-        </tr>
-        <tr>
-            <td class="tg-lap0">Stadt:</td>
-            <td class="tg-lap0"><input th:name="city" type="text" th:placeholder="${user.address.city}"/></td>
-        </tr>
-        <tr>
-            <td class="tg-lap0">Land:</td>
-            <td class="tg-lap0"><input th:name="country" type="text" th:placeholder="${user.address.country}"/></td>
-        </tr>
-        <tr>
-            <td class="tg-lap0">Hausnummer:</td>
-            <td class="tg-lap0"><input th:name="houseNumber" type="text" th:placeholder="${user.address.houseNumber}"/></td>
-        </tr>
-        <tr>
-            <td class="tg-lap0">Postleitzahl:</td>
-            <td class="tg-lap0"><input th:name="postCode" type="text" th:placeholder="${user.address.postCode}"/></td>
-        </tr>
-    </table>
-    <td><button type="submit" class="btn btn-primary">Aktuallisieren</button></td>
-    <td><a th:href="|/|" > <button class="btn btn-warning" type="button">Abbrechen</button></a></td>
-</form>
-</body>
+<html lang="de" xmlns:th="http://www.thymeleaf.org" layout:decorate="~{Admin/admin_overview}"
+      xmlns:layout="http://www.w3.org/1999/xhtml">
+
+    <head>
+        <title>Benutzerverwaltung - Rhinoshare</title>
+    </head>
+
+    <body>
+        <div class="container" layout:fragment="content">
+            <div>
+                <a href="/admin/usermanagement">
+                    <button class="button_std">Zurück</button>
+                </a>
+            </div>
+            <form method="post" style="text-align: center;">
+                <table class="tg">
+                    <tr>
+                        <td class="tg-lap0">Username:</td>
+                        <td class="tg-lap0"><input th:name="username" type="text" th:placeholder="${user.username}"/></td>
+                    </tr>
+                    <tr>
+                        <td class="tg-lap0">Nachname:</td>
+                        <td class="tg-lap0"><input th:name="name" type="text" th:placeholder="${user.name}"/></td>
+                    </tr>
+                    <tr>
+                        <td class="tg-lap0">Vorname:</td>
+                        <td class="tg-lap0"><input th:name="surname" type="text" th:placeholder="${user.surname}"/></td>
+                    </tr>
+                    <tr>
+                        <td class="tg-lap0">E-Mail Adresse:</td>
+                        <td class="tg-lap0"><input th:name="email" type="email" th:placeholder="${user.email}"/></td>
+                    </tr>
+                    <tr>
+                        <td class="tg-lap0">Straße:</td>
+                        <td class="tg-lap0"><input th:name="street" type="text" th:placeholder="${user.address.street}"/></td>
+                    </tr>
+                    <tr>
+                        <td class="tg-lap0">Stadt:</td>
+                        <td class="tg-lap0"><input th:name="city" type="text" th:placeholder="${user.address.city}"/></td>
+                    </tr>
+                    <tr>
+                        <td class="tg-lap0">Land:</td>
+                        <td class="tg-lap0"><input th:name="country" type="text" th:placeholder="${user.address.country}"/></td>
+                    </tr>
+                    <tr>
+                        <td class="tg-lap0">Hausnummer:</td>
+                        <td class="tg-lap0"><input th:name="houseNumber" type="text" th:placeholder="${user.address.houseNumber}"/></td>
+                    </tr>
+                    <tr>
+                        <td class="tg-lap0">Postleitzahl:</td>
+                        <td class="tg-lap0"><input th:name="postCode" type="text" th:placeholder="${user.address.postCode}"/></td>
+                    </tr>
+                </table>
+                <td><button type="submit" class="btn btn-primary">Aktuallisieren</button></td>
+                <td><a th:href="|/|" > <button class="btn btn-warning" type="button">Abbrechen</button></a></td>
+            </form>
+        </div>
+    </body>
+
 </html>
