+++ RhinoshareApp/rhinoshareapp/src/main/resources/templates/mainpage.html
@@ -1,5 +1,6 @@
 <!DOCTYPE html>
 <html lang="de" xmlns:th="http://www.thymeleaf.org" xmlns:layout="http://www.w3.org/1999/xhtml">
+
     <head>
         <meta charset="utf-8">
         <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
@@ -12,22 +13,19 @@
         <img id="titleImage" src="/img/rhinoLogoBig_share.png" width="300" height="75">
 
         <ol class="navigationbar">
-            <li class="navbaritem"><a class="active" href="/">Rhino-Übersicht</a></li>
-            <li class="navbaritem"><a href="/article/">Artikel</a></li>
+            <li class="navbaritem" th:classappend="${mainActive}"><a href="/">Rhino-Übersicht</a></li>
+            <li class="navbaritem" th:classappend="${articleActive}"><a href="/article/">Artikel</a></li>
 
-            <li class="navbaritem"><a th:href="|/${username}|">Ausleihen</a></li>
-
-
-            <li style="float: right;" class="navbaritem" href th:if="${loggedIn == 'false'}">
+            <li style="float: right;" class="navbaritem" th:classappend="${userActive}" href th:if="${loggedIn == 'false'}">
                 <a href="/login">Einloggen</a>
             </li>
-            <li style="float: right" class="navbaritem" href th:if="${loggedIn == 'true'}">
+            <li style="float: right" class="navbaritem" th:classappend="${userActive}" href th:if="${loggedIn == 'true'}">
                 <a href="/logout">Ausloggen</a>
             </li>
-            <li style="float: right" class="navbaritem" href th:if="${loggedIn == 'false'}">
+            <li style="float: right" class="navbaritem" th:classappend="${userActive}" href th:if="${loggedIn == 'false'}">
                 <a href="/newaccount">Registrieren</a>
             </li>
-            <li style="float: right" class="navbaritem" href th:if="${loggedIn == 'true'}">
+            <li style="float: right" class="navbaritem"  th:classappend="${userActive}"href th:if="${loggedIn == 'true'}">
                 <div class="dropdown">
                     <a href="/edit" th:text="${surname}" class="dropbtn">Benutzer</a>
                     <div class="dropdown-content">
@@ -56,4 +54,5 @@
             Content goes here
         </div>
     </div>
+
 </html>
